feat: implement controllers and JWT authentication, update DTOs and services

+++ ExpenseTracker/Services/CategoryService.cs
@@ -1,5 +1,6 @@
 using ExpenseTracker.Data;
 using ExpenseTracker.DTO;
+using ExpenseTracker.Exceptions;
 using ExpenseTracker.Models;
 
 namespace ExpenseTracker.Services;
@@ -13,81 +14,64 @@ public class CategoryService : ICategoryService
         _context = context;
     }
     
-    public List<CategoryResponseDto> GetAllCategories()
+    public List<CategoryResponseDto> GetAllCategories(int  userId)
     {
-        var categories = _context.Categories.ToList();
-        var categoriesDto = categories.Select(c=> new CategoryResponseDto
-        {
-            Id = c.Id,
-            Name = c.Name,
-            Description = c.Description,
-        }).ToList();
-        
-        return categoriesDto;
+        return _context.Categories
+            .Where(c => c.UserId == userId || c.UserId == null)
+            .Select(c=> new CategoryResponseDto
+            {
+                Id = c.Id,
+                Name = c.Name,
+                Description = c.Description,
+            })
+            .ToList();
     }
 
-    public ServiceResult<CategoryResponseDto> CreateCategory(CreateCategoryRequestDto request)
+    public CategoryResponseDto CreateCategory(int userId, CreateCategoryRequestDto request)
     {
         var normalizedName = request.Name.Trim().ToLower();
-        var exists = _context.Categories.Any(c=> c.Name.ToLower() == normalizedName);
+        var exists = _context.Categories.Any(c=> c.Name.ToLower() == normalizedName&&(c.UserId == userId || c.UserId == null));
 
         if (exists)
-        {
-            return new ServiceResult<CategoryResponseDto>
-            {
-                Success = false,
-                Message = $"Категория с названием {request.Name} уже существует."
-            };
-        }
+            throw new BusinessException("Такая категория уже существует");
 
         var name = request.Name.Trim();
         
         var category = new Category(
-            null,
+            userId,
             name,
             request.Description);
         
         _context.Categories.Add(category);
         _context.SaveChanges();
 
-        return new ServiceResult<CategoryResponseDto>
+        return new CategoryResponseDto
         {
-            Success = true,
-            Data = new CategoryResponseDto
-            {
-                Id = category.Id,
-                Name = category.Name,
-                Description = category.Description,
-            }
+            Id = category.Id,
+            Name = category.Name,
+            Description = category.Description,
         };
     }
 
-    public ServiceResult<CategoryResponseDto> UpdateCategory(int categoryId, UpdateCategoryRequestDto request)
+    public CategoryResponseDto UpdateCategory(int userId, int categoryId, UpdateCategoryRequestDto request)
     {
-        var category = _context.Categories.FirstOrDefault(c => c.Id == categoryId);
+        var category = _context.Categories.FirstOrDefault(c => c.Id == categoryId && (c.UserId == userId || c.UserId == null));
         if (category == null)
-        {
-            return new ServiceResult<CategoryResponseDto>
-            {
-                Success = false,
-                Message = $"Категории с айди: {categoryId} не существует."
-            };
-        }
+            throw new BusinessException("Такой категории не существует");
+        
+        if (category.UserId != userId)
+            throw new BusinessException("Категория является системной, ее нельзя поменять");
 
         if (!string.IsNullOrWhiteSpace(request.Name))
         {
             var newName = request.Name.Trim();
             if (newName != category.Name)
             {
-                var exists = _context.Categories.Any(c=> c.Name.ToLower() == newName.ToLower());
+                var exists = _context.Categories.Any(c=> c.Name.ToLower() == newName.ToLower() && (c.UserId == userId || c.UserId == null));
 
                 if (exists)
                 {
-                    return new ServiceResult<CategoryResponseDto>
-                    {
-                        Success = false,
-                        Message = $"Категория с названием {newName} уже существует."
-                    };
+                    throw new BusinessException("Категория с таким именем уже существует");
                 }
                 
                 category.Name = newName;
@@ -95,44 +79,32 @@ public class CategoryService : ICategoryService
         }
 
         if (request.Description != null)
-        {
             category.Description = request.Description;
-        }
         
         _context.SaveChanges();
 
-        return new ServiceResult<CategoryResponseDto>
+        return new CategoryResponseDto
         {
-            Success = true,
-            Data = new CategoryResponseDto
-            {
-                Id = category.Id,
-                Name = category.Name,
-                Description = category.Description,
-            }
+            Id = category.Id,
+            Name = category.Name,
+            Description = category.Description,
         };
     }
 
-    public ServiceResult<bool> DeleteCategory(int id)
+    public void DeleteCategory(int userId, int id)
     {
         var category = _context.Categories.FirstOrDefault(c => c.Id == id);
 
         if (category == null)
-        {
-            return new ServiceResult<bool>
-            {
-                Success = false,
-                Message = $"Категории с айди {id} не существует."
-            };
-        }
+            throw new BusinessException("Категории с таким айди нет");
+
+        if (category.UserId == null)
+            throw new BusinessException("Вы не можете удалить системную категорию");
+        
+        if (category.UserId != userId)
+            throw new BusinessException("Вы не можете удалить чужую категорию");
         
         _context.Categories.Remove(category);
         _context.SaveChanges();
-
-        return new ServiceResult<bool>
-        {
-            Success = true,
-            Message = "Категория успешно удалена."
-        };
     }
 }
